Add to cart created and change the ProductModal
