+++ client-frontend/src/security/api/CartApi.js
@@ -1,0 +1,12 @@
+import { api } from '../ApiClient'
+
+export const addOrUpdateCartItem = (productId, quantity) => {
+
+    return api.put(`/my-cart`, {}, {
+        params: {
+            productId: productId,
+            quantity: quantity
+        }
+    });
+
+}
